+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version 1.0.0-beta7
+
+## Bugfixes
+
+* None
+
+## Features
+
+* Move getRowStoreId() method from AbstractEavSubject to AbstractSubjectMethod
+* Move getRowStoreId() method from AbstractObserverTrait to AbstractObserver
+
 # Version 1.0.0-beta6
 
 ## Bugfixes